- Merge avec refactor router et modification page d'accueil
