+++ src/views/AppMain.vue
@@ -4,32 +4,32 @@
       <div class="py-5"/>
       <v-row>
         <v-col class="text-center" cols="12" md="6">
-          <bouton-item icon="mdi-percent" route="tauxRecouv">
+          <bouton-item icon="mdi-percent" route="recouvrement-etape-1-selection-rcr">
             Calculer mon taux de recouvrement
           </bouton-item>
         </v-col>
         <v-col class="text-center" cols="12" md="6">
-          <bouton-item icon="mdi-chart-bar" route="recouvrementTableau">
+          <bouton-item icon="mdi-chart-bar" route="recouvrement-tableau">
             Gérer mes taux de recouvrement
           </bouton-item>
         </v-col>
         <v-col class="text-center" cols="12" md="6">
-          <bouton-item icon="mdi-plus" route="nouveauxExemplaires">
+          <bouton-item icon="mdi-plus" route="exemplarisation-etape-1-selection-rcr">
             Créer des exemplaires
           </bouton-item>
         </v-col>
         <v-col class="text-center" cols="12" md="6">
-          <bouton-item icon="mdi-chart-bar" route="exemplarisationTableau">
+          <bouton-item icon="mdi-chart-bar" route="exemplarisation-tableau">
             Gérer mes créations
           </bouton-item>
         </v-col>
         <v-col class="text-center" cols="12" md="6">
-          <bouton-item icon="mdi-pencil" route="modifExempl">
+          <bouton-item icon="mdi-pencil" route="modification-etape-1-selection-rcr">
             Modifier des exemplaires
           </bouton-item>
         </v-col>
         <v-col class="text-center" cols="12" md="6">
-          <bouton-item icon="mdi-chart-bar" route="modificationTableau">
+          <bouton-item icon="mdi-chart-bar" route="modification-tableau">
             Gérer mes modifications
           </bouton-item>
         </v-col>
